% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertR.R
\name{convertR}
\alias{convertR}
\title{Convert Excel Files to CSV (R version)}
\usage{
convertR(dir_path)
}
\arguments{
\item{dir_path}{Character. Directory containing files.}
}
\value{
Invisibly returns TRUE.
}
\description{
Lists files in a directory and converts `.xls` and `.xlsx` files to `.csv`.
}
\examples{
\dontrun{
convertR("data/")
}
}
