% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myssxy.R
\name{myssxy}
\alias{myssxy}
\title{Calculation of Cross Product}
\usage{
myssxy(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a scalar (the cross product)
}
\description{
Calculation of Cross Product
}
\examples{
myssxy(x = 1:10, y = 1:10)
}
