% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Directory Explorer App}
\usage{
run_app(dir_path = fs::path_home(), top_n = 10, hash = FALSE)
}
\arguments{
\item{dir_path}{Starting directory (default = home directory)}

\item{top_n}{Default number of top file types}

\item{hash}{Whether duplicate detection is enabled by default}
}
\description{
Launch the Shiny app with optional arguments.
}
