% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myvar.R
\name{myvar}
\alias{myvar}
\title{Calculation of sample variance}
\usage{
myvar(y)
}
\arguments{
\item{y}{a numeric vector}
}
\value{
a scalar (variance)
}
\description{
Calculation of sample variance
}
\examples{
myvar(1:10)
\dontrun{myvar(1:10)}
}
